Routing debut
Took the body of the html out, only templates now.
The basic (layout) template for the whole site has the navbar and the sidebar inside it.
The mainpage (or homepage) should always have the ID main !!!! and is for the moment the sum of the old templates into two templates :
- first the welcome words and 5 choices
- second the news feed which will be separeted and will need the infinite scroll