Infinite scroll
Put in place the infinite scroll on all pages with topics
In place for the single topics and comments as well, but not yet working.
Had to change the layout of the main Id to have the entire window to scroll (window) is a parameter of the infinite scroll.
Issue with the userId and createdBy that has arrised with all pages.