SearchTool
I repaired the search tool :
- I managed to use the $or query inside my find query, returning either a matching result in category, description or title
- I also added another helper to the search template, in order to responsively count the number of results and display it on the template.
It is basically a function (found wih the handlebars on the search template in the html).
It uses the exact same reactive function than for the search query except instead of returning topics it returns a number using find().count(), it then displays either the number if it is above 0, and if it is 0 it returns no to be consistant in the template :
either : "You have X result!"
or: "You have no result!"
I did this using an if statement
Beware, I did have an issue, it was counting 8 instead of X>8 since it was limited with my session.set topic limit..., so when removing that constraint it finally worked.
Hurray!

SearchTool DONE ! :D